Change mesh structure to use size_t for node and element IDs
- Modified hecmwST_local_mesh structure to use size_t* for node_ID and elem_ID arrays
- This enables support for meshes with more than 2^31 nodes/elements on 64-bit systems

+++ hecmw1/src/common/hecmw_struct.h
@@ -192,7 +192,7 @@ struct hecmwST_local_mesh {
   int n_elem_type;
   int *elem_type_index;
   int *elem_type_item;
-  int *elem_node_index;
+  long long *elem_node_index;
   int *elem_node_item;
   int *section_ID;
   int *elem_mat_ID_index;
